requirements.txt
transformers==4.40.0
torch
pillow
pinecone
pytest
faiss-cpu
dvc
python-dotenv
httpx==0.27.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
Pillow==9.5.0
loguru==0.7.0
fastapi==0.96.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
prometheus-client==0.17.1
prometheus-fastapi-instrumentator
python-json-logger